move confs
